+++ lib/flutter_storyboard.dart
@@ -1,13 +1,24 @@
 library storyboard;
 
+import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
+
+import 'src/custom_rect_clipper.dart';
+import 'src/nested_app.dart';
+import 'src/rounded_label.dart';
 
 const _kSpacing = 40.0;
 
 class StoryBoard extends StatefulWidget {
   /// Wrap your Material App with this widget
-  final MaterialApp child;
+  final MaterialApp _materialApp;
+
+  /// Wrap your Widgets App with this widget
+  final WidgetsApp _widgetsApp;
+
+  /// Wrap your Cupertino App with this widget
+  final CupertinoApp _cupertinoApp;
 
   /// Size for each screen
   final Size screenSize;
@@ -30,9 +41,17 @@ class StoryBoard extends StatefulWidget {
   /// Override the default app bar of the storyboard
   final AppBar customAppBar;
 
+  /// Custom screens you want to show
+  final List<Widget> customScreens;
+
+  /// Custom routes you want to show with test data
+  final List<RouteSettings> customRoutes;
+
   const StoryBoard({
     Key key,
-    @required this.child,
+    MaterialApp materialApp,
+    WidgetsApp widgetsApp,
+    CupertinoApp cupertinoApp,
     this.enabled = true,
     this.screenSize = const Size(400, 700),
     this.initialOffset,
@@ -40,7 +59,68 @@ class StoryBoard extends StatefulWidget {
     this.initialScale,
     this.scaleChanged,
     this.customAppBar,
-  }) : super(key: key);
+    this.customRoutes,
+    this.customScreens,
+  })  : _materialApp = materialApp,
+        _widgetsApp = widgetsApp,
+        _cupertinoApp = cupertinoApp,
+        assert(
+            materialApp != null || widgetsApp != null || cupertinoApp != null),
+        super(key: key);
+
+  const StoryBoard.material({
+    Key key,
+    Widget child,
+    this.enabled = true,
+    this.screenSize = const Size(400, 700),
+    this.initialOffset,
+    this.offsetChanged,
+    this.initialScale,
+    this.scaleChanged,
+    this.customAppBar,
+    this.customRoutes,
+    this.customScreens,
+  })  : _materialApp = child,
+        _widgetsApp = null,
+        _cupertinoApp = null,
+        assert(child != null),
+        super(key: key);
+
+  const StoryBoard.custom({
+    Key key,
+    Widget child,
+    this.enabled = true,
+    this.screenSize = const Size(400, 700),
+    this.initialOffset,
+    this.offsetChanged,
+    this.initialScale,
+    this.scaleChanged,
+    this.customAppBar,
+    this.customRoutes,
+    this.customScreens,
+  })  : _materialApp = null,
+        _widgetsApp = child,
+        _cupertinoApp = null,
+        assert(child != null),
+        super(key: key);
+
+  const StoryBoard.cupertino({
+    Key key,
+    Widget child,
+    this.enabled = true,
+    this.screenSize = const Size(400, 700),
+    this.initialOffset,
+    this.offsetChanged,
+    this.initialScale,
+    this.scaleChanged,
+    this.customAppBar,
+    this.customRoutes,
+    this.customScreens,
+  })  : _materialApp = null,
+        _widgetsApp = null,
+        _cupertinoApp = child,
+        assert(child != null),
+        super(key: key);
 
   @override
   StoryboardController createState() => StoryboardController();
@@ -50,6 +130,38 @@ class StoryboardController extends State<StoryBoard> {
   double _scale;
   Offset _offset;
   FocusNode _focusNode;
+
+  Widget get app {
+    if (materialApp != null) return materialApp;
+    if (widgetsApp != null) return widgetsApp;
+    if (cupertinoApp != null) return cupertinoApp;
+    return null;
+  }
+
+  MaterialApp get materialApp => widget?._materialApp;
+  WidgetsApp get widgetsApp => widget?._widgetsApp;
+  CupertinoApp get cupertinoApp => widget?._cupertinoApp;
+
+  Widget get home {
+    if (materialApp != null) return materialApp?.home;
+    if (widgetsApp != null) return widgetsApp?.home;
+    if (cupertinoApp != null) return cupertinoApp?.home;
+    return null;
+  }
+
+  Map<String, WidgetBuilder> get routes {
+    if (materialApp != null) return materialApp?.routes;
+    if (widgetsApp != null) return widgetsApp?.routes;
+    if (cupertinoApp != null) return cupertinoApp?.routes;
+    return null;
+  }
+
+  String get initialRoute {
+    if (materialApp != null) return materialApp?.initialRoute;
+    if (widgetsApp != null) return widgetsApp?.initialRoute;
+    if (cupertinoApp != null) return cupertinoApp?.initialRoute;
+    return null;
+  }
 
   @override
   void initState() {
@@ -63,7 +175,13 @@ class StoryboardController extends State<StoryBoard> {
 
   @override
   void didUpdateWidget(StoryBoard oldWidget) {
-    if (oldWidget.child != widget.child) {
+    if (oldWidget._widgetsApp != widget._widgetsApp) {
+      if (mounted) setState(() {});
+    }
+    if (oldWidget._materialApp != widget._materialApp) {
+      if (mounted) setState(() {});
+    }
+    if (oldWidget._cupertinoApp != widget._cupertinoApp) {
       if (mounted) setState(() {});
     }
     super.didUpdateWidget(oldWidget);
@@ -77,14 +195,14 @@ class StoryboardController extends State<StoryBoard> {
 
   @override
   Widget build(BuildContext context) {
-    final base = widget.child;
     final _size = widget.screenSize;
-    if (!widget.enabled) return base;
+    _row = 0;
+    if (!widget.enabled) return app;
     return MaterialApp(
-      debugShowCheckedModeBanner: base.debugShowCheckedModeBanner,
-      themeMode: base.themeMode,
-      theme: base.theme,
-      darkTheme: base.darkTheme,
+      debugShowCheckedModeBanner: false,
+      themeMode: materialApp?.themeMode ?? ThemeMode.system,
+      theme: materialApp?.theme ?? ThemeData.light(),
+      darkTheme: materialApp?.darkTheme ?? ThemeData.dark(),
       home: Scaffold(
         appBar: widget?.customAppBar ??
             AppBar(
@@ -119,31 +237,67 @@ class StoryboardController extends State<StoryBoard> {
                 child: Stack(
                   fit: StackFit.expand,
                   children: [
-                    if (base?.home != null)
-                      _addChild(
-                        base,
-                        base.home,
-                        Offset(0, 10),
-                        'Home',
-                      ),
-                    if (base?.initialRoute != null)
-                      _addChild(
-                        base,
-                        base.routes[base.initialRoute](context),
-                        Offset(base?.home != null ? _size.width + _kSpacing : 0,
-                            10),
-                        'Initial Route',
-                      ),
-                    if (base?.routes != null) ...[
-                      for (var r = 0; r < base.routes.keys.length; r++)
-                        _addChild(
-                          base,
-                          base.routes[base.routes.keys.toList()[r]](context),
-                          Offset((_size.width + _kSpacing) * r,
-                              (_size.height + _kSpacing) + 40),
-                          base.routes.keys.toList()[r],
-                        ),
-                    ]
+                    ..._addRow(
+                        0,
+                        (row) => [
+                              if (home != null)
+                                _addChild(
+                                  null,
+                                  offset: Offset(0, 10),
+                                  label: 'Home',
+                                  child: home,
+                                ),
+                              if (initialRoute != null)
+                                _addChild(
+                                  RouteSettings(name: initialRoute),
+                                  offset: Offset(
+                                      home != null
+                                          ? ((_size.width + _kSpacing) * 1)
+                                          : 0,
+                                      10),
+                                  label: 'Initial Route',
+                                ),
+                            ]),
+                    if (routes != const <String, WidgetBuilder>{})
+                      ..._addRow(
+                          1,
+                          (row) => [
+                                for (var i = 0; i < routes.keys.length; i++)
+                                  _addChild(
+                                    RouteSettings(
+                                        name: routes.keys.toList()[i]),
+                                    offset: _getOffset(_size, row, i),
+                                    label: routes.keys.toList()[i],
+                                  ),
+                              ]),
+                    if (widget.customScreens != null)
+                      ..._addRow(
+                          2,
+                          (row) => [
+                                for (var i = 0;
+                                    i < widget.customScreens.length;
+                                    i++)
+                                  _addChild(
+                                    null,
+                                    offset: _getOffset(_size, row, i),
+                                    label: widget.customScreens[i].toString(),
+                                    child: widget.customScreens[i],
+                                    customWidget: true,
+                                  ),
+                              ]),
+                    if (widget.customRoutes != null)
+                      ..._addRow(
+                          3,
+                          (row) => [
+                                for (var i = 0;
+                                    i < widget.customRoutes.length;
+                                    i++)
+                                  _addChild(
+                                    widget.customRoutes[i],
+                                    offset: _getOffset(_size, row, i),
+                                    label: widget.customRoutes[i].name,
+                                  ),
+                              ]),
                   ],
                 ),
               ),
@@ -152,6 +306,11 @@ class StoryboardController extends State<StoryBoard> {
         ),
       ),
     );
+  }
+
+  Offset _getOffset(Size _size, int row, [int i = 1]) {
+    return Offset((_size.width + _kSpacing) * i,
+        (((_size.height + _kSpacing + 40) * row)));
   }
 
   void _handleKeyPress(RawKeyEvent key) {
@@ -197,11 +356,19 @@ class StoryboardController extends State<StoryBoard> {
     }
   }
 
-  Positioned _addChild(MaterialApp base, Widget child,
-      [Offset offset = Offset.zero, String label]) {
+  int _row = 0;
+  List<Widget> _addRow(int place, List<Widget> Function(int) children) =>
+      children(_row++);
+
+  Positioned _addChild(
+    RouteSettings route, {
+    Offset offset = Offset.zero,
+    String label,
+    Widget child,
+    bool customWidget = false,
+  }) {
     final _top = _offset.dy + (_scale * offset.dy);
     final _left = _offset.dx + (_scale * offset.dx);
-    final base = widget.child;
     final _size = widget.screenSize;
     return Positioned(
       top: _top,
@@ -211,61 +378,29 @@ class StoryboardController extends State<StoryBoard> {
         child: Column(
           mainAxisSize: MainAxisSize.min,
           children: [
-            SizedBox(
-              width: _size.width,
-              height: _size.height,
+            SizedBox.fromSize(
+              size: _size,
               child: Material(
                 elevation: 4,
                 child: ClipRect(
                   clipper: CustomRect(Offset(0, 0)),
-                  child: MaterialApp(
-                    debugShowCheckedModeBanner: false,
-                    routes: base.routes,
-                    onGenerateRoute: base.onGenerateRoute,
-                    onGenerateInitialRoutes: base.onGenerateInitialRoutes,
-                    onUnknownRoute: base.onUnknownRoute,
-                    home: child,
+                  child: NestedApp(
+                    route: route,
+                    child: child,
+                    size: _size,
+                    customWidget: customWidget,
+                    materialApp: materialApp,
+                    cupertinoApp: cupertinoApp,
+                    widgetsApp: widgetsApp,
+                    routes: routes,
                   ),
                 ),
               ),
             ),
-            if (label != null) Center(child: _addLabel(label)),
+            if (label != null) Center(child: RoundedLabel(label)),
           ],
         ),
       ),
     );
-  }
-
-  Widget _addLabel(String label) {
-    return Container(
-      margin: EdgeInsetsDirectional.only(top: 10),
-      padding: EdgeInsets.symmetric(vertical: 6, horizontal: 12),
-      decoration: BoxDecoration(
-        color: Colors.black,
-        borderRadius: BorderRadius.circular(20),
-      ),
-      child: Text(
-        label,
-        style: TextStyle(
-          color: Colors.white,
-          fontSize: 20,
-        ),
-      ),
-    );
-  }
-}
-
-class CustomRect extends CustomClipper<Rect> {
-  final Offset offset;
-
-  CustomRect(this.offset);
-  @override
-  Rect getClip(Size size) {
-    return Rect.fromLTWH(offset.dx, offset.dy, size.width, size.height);
-  }
-
-  @override
-  bool shouldReclip(CustomRect oldClipper) {
-    return true;
   }
 }