+++ lib/src/rounded_label.dart
@@ -1,0 +1,25 @@
+import 'package:flutter/material.dart';
+
+class RoundedLabel extends StatelessWidget {
+  final String label;
+
+  const RoundedLabel(this.label);
+  @override
+  Widget build(BuildContext context) {
+    return Container(
+      margin: EdgeInsetsDirectional.only(top: 10),
+      padding: EdgeInsets.symmetric(vertical: 6, horizontal: 12),
+      decoration: BoxDecoration(
+        color: Colors.black,
+        borderRadius: BorderRadius.circular(20),
+      ),
+      child: Text(
+        label ?? '',
+        style: TextStyle(
+          color: Colors.white,
+          fontSize: 20,
+        ),
+      ),
+    );
+  }
+}