+++ CHANGELOG.md
@@ -1,3 +1,12 @@
-## [0.0.1] - TODO: Add release date.
+## 0.0.2
 
-* TODO: Describe initial release.
+* Adding onGenerateRoute, onUnknownRoute, initialRoute
+* Added Support for Custom Widgets/Screens
+* Updated Documentation
+* Updated Example
+* Added Support for Keyboard Controls for Zoom/Pan @Nash0x7E2
+* Added Constraint on Flutter v1.14.7
+
+## 0.0.1
+
+* Support for Named Routes and Home Field